Move files to represent more common go package structure.
