+++ .gitignore
@@ -8,4 +8,3 @@
 /demo/
 /github/
 /node_modules/
-/pkg/
