+++ cmd/esbuild/version.go
@@ -1,0 +1,3 @@
+package main
+
+const esbuildVersion = "0.0.17"
