+++ npm/esbuild-linux-64/package.json
@@ -1,6 +1,6 @@
 {
   "name": "esbuild-linux-64",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "The Linux 64-bit binary for esbuild, a JavaScript bundler.",
   "repository": "https://github.com/evanw/esbuild",
   "license": "MIT",
