+++ npm/esbuild-wasm/package.json
@@ -1,6 +1,6 @@
 {
   "name": "esbuild-wasm",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "The cross-platform WebAssembly binary for esbuild, a JavaScript bundler.",
   "repository": "https://github.com/evanw/esbuild",
   "license": "MIT",
