+++ npm/esbuild-windows-64/package.json
@@ -1,6 +1,6 @@
 {
   "name": "esbuild-windows-64",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "The Windows 64-bit binary for esbuild, a JavaScript bundler.",
   "repository": "https://github.com/evanw/esbuild",
   "license": "MIT",
