+++ npm/esbuild/package.json
@@ -1,6 +1,6 @@
 {
   "name": "esbuild",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "An extremely fast JavaScript bundler and minifier.",
   "repository": "https://github.com/evanw/esbuild",
   "scripts": {
