+++ pkg/bundler/bundler.go
@@ -4,13 +4,13 @@ import (
 	"bytes"
 	"encoding/base64"
 	"fmt"
-	"github.com/evanw/esbuild/src/esbuild/ast"
-	"github.com/evanw/esbuild/src/esbuild/fs"
-	"github.com/evanw/esbuild/src/esbuild/lexer"
-	"github.com/evanw/esbuild/src/esbuild/logging"
-	"github.com/evanw/esbuild/src/esbuild/parser"
-	"github.com/evanw/esbuild/src/esbuild/printer"
-	"github.com/evanw/esbuild/src/esbuild/resolver"
+	"github.com/evanw/esbuild/pkg/ast"
+	"github.com/evanw/esbuild/pkg/fs"
+	"github.com/evanw/esbuild/pkg/lexer"
+	"github.com/evanw/esbuild/pkg/logging"
+	"github.com/evanw/esbuild/pkg/parser"
+	"github.com/evanw/esbuild/pkg/printer"
+	"github.com/evanw/esbuild/pkg/resolver"
 	"sort"
 	"strings"
 	"sync"
