+++ pkg/lexer/lexer.go
@@ -15,8 +15,8 @@ package lexer
 
 import (
 	"fmt"
-	"github.com/evanw/esbuild/src/esbuild/ast"
-	"github.com/evanw/esbuild/src/esbuild/logging"
+	"github.com/evanw/esbuild/pkg/ast"
+	"github.com/evanw/esbuild/pkg/logging"
 	"strconv"
 	"strings"
 	"unicode"
