+++ pkg/lexer/tables.go
