+++ pkg/logging/logging.go
@@ -7,7 +7,7 @@ package logging
 
 import (
 	"fmt"
-	"github.com/evanw/esbuild/src/esbuild/ast"
+	"github.com/evanw/esbuild/pkg/ast"
 	"os"
 	"os/exec"
 	"runtime"
