--- src/esbuild/main/version.go
@@ -1,3 +1,0 @@
-package main
-
-const esbuildVersion = "0.0.16"
